requirements.txt
asgiref==3.9.0
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
